Refactor: split credential handlers, extract helper, unify disable path
- Move credential binding handlers (assign/remove/get) to a new
  api/agent_connector_credentials.go file — keeps agent_connectors.go
  focused on connector lifecycle and avoids a single megafile
- Extract deleteCredentialAndVault helper: cleaner, testable unit that
  treats NotFound as a no-op (idempotent delete)
- Unify disable handler code paths: always use BeginOrContinue so both
  paths benefit from the same transaction semantics; single RespondJSON
  call at the end eliminates the duplicated response block
- Use t.Context() instead of context.Background() in new test for
  consistent cancellation propagation
